# openapi.yaml
openapi: 3.0.3
info:
  title: Stellar Footprint Service API
  description: |
    Pre-flight simulation and footprint extraction for Soroban smart contract transactions.

    This service automates the complex footprint extraction process — a required step before
    submitting any Soroban transaction to the Stellar network.
  version: 1.0.0
  license:
    name: MIT

servers:
  - url: /api/v1
    description: Current API version

tags:
  - name: simulation
    description: Transaction simulation and footprint extraction
  - name: utilities
    description: XDR decoding and health checks

paths:
  /health:
    get:
      tags: [utilities]
      summary: Liveness check
      description: Returns service health status. Used by load balancers and uptime monitors.
      operationId: getHealth
      responses:
        "200":
          description: Service is healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"
              example:
                status: ok
                uptime: 123.45
                version: 1.0.0
                timestamp: "2024-01-01T00:00:00.000Z"

  /simulate:
    post:
      tags: [simulation]
      summary: Simulate a Soroban transaction
      description: |
        Submits a transaction XDR to the Stellar RPC `simulateTransaction` endpoint,
        extracts the footprint (read-only and read-write ledger keys), and returns
        resource cost estimates.
      operationId: simulate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SimulateRequest"
            example:
              xdr: "AAAAAgAAAAC..."
              network: testnet
      responses:
        "200":
          description: Simulation succeeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SimulateSuccessResponse"
              example:
                success: true
                footprint:
                  readOnly: ["AAAABgAAAAHZ...", "AAAABgAAAAHa..."]
                  readWrite: ["AAAABgAAAAHb..."]
                cost:
                  cpuInsns: "1234567"
                  memBytes: "8192"
        "400":
          description: Missing or invalid request body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: "Missing required field: xdr"
        "422":
          description: Simulation failed (contract error or restoration required)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SimulateFailureResponse"
              examples:
                contractError:
                  summary: Contract execution error
                  value:
                    success: false
                    error: "contract panic: assertion failed"
                restoreRequired:
                  summary: Ledger entry restoration required
                  value:
                    success: false
                    error: "Transaction requires ledger entry restoration before simulation."
        "429":
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: "Too many requests, please try again later."
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: "Unexpected error"

  /decode:
    get:
      tags: [utilities]
      summary: Decode XDR to human-readable JSON
      description: |
        Decodes a base64-encoded XDR string into a human-readable JSON representation.
        Useful for debugging transaction structures and ledger entries.
      operationId: decodeXdr
      parameters:
        - name: xdr
          in: query
          required: true
          description: Base64-encoded XDR string to decode
          schema:
            type: string
            example: "AAAAAgAAAAC..."
        - name: type
          in: query
          required: false
          description: |
            XDR type hint for decoding. If omitted the service will attempt
            auto-detection. Common values: `transaction`, `ledgerEntry`, `operationResult`.
          schema:
            type: string
            example: transaction
      responses:
        "200":
          description: XDR decoded successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DecodeResponse"
              example:
                type: transaction
                decoded:
                  fee: 100
                  operations: []
        "400":
          description: Missing or invalid XDR parameter
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: "Missing required query parameter: xdr"
        "422":
          description: XDR could not be decoded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: "Failed to decode XDR: invalid base64"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    HealthResponse:
      type: object
      required: [status, uptime, timestamp]
      properties:
        status:
          type: string
          enum: [ok]
        uptime:
          type: number
          description: Process uptime in seconds
        version:
          type: string
          description: Service version from package.json
        timestamp:
          type: string
          format: date-time

    SimulateRequest:
      type: object
      required: [xdr]
      properties:
        xdr:
          type: string
          description: Base64-encoded Soroban transaction XDR
          example: "AAAAAgAAAAC..."
        network:
          type: string
          enum: [testnet, mainnet]
          default: testnet
          description: Stellar network to simulate against
        ledgerSequence:
          type: integer
          description: |
            Optional ledger sequence to simulate against.
            Useful for reproducing historical simulation results.
          example: 12345678

    Footprint:
      type: object
      required: [readOnly, readWrite]
      properties:
        readOnly:
          type: array
          items:
            type: string
          description: Base64-encoded read-only ledger keys
        readWrite:
          type: array
          items:
            type: string
          description: Base64-encoded read-write ledger keys

    ResourceCost:
      type: object
      required: [cpuInsns, memBytes]
      properties:
        cpuInsns:
          type: string
          description: CPU instructions consumed (as string to avoid precision loss)
          example: "1234567"
        memBytes:
          type: string
          description: Memory bytes consumed (as string to avoid precision loss)
          example: "8192"

    SimulateSuccessResponse:
      type: object
      required: [success, footprint, cost]
      properties:
        success:
          type: boolean
          enum: [true]
        footprint:
          $ref: "#/components/schemas/Footprint"
        cost:
          $ref: "#/components/schemas/ResourceCost"

    SimulateFailureResponse:
      type: object
      required: [success, error]
      properties:
        success:
          type: boolean
          enum: [false]
        error:
          type: string
          description: Human-readable error message from the simulation

    DecodeResponse:
      type: object
      required: [type, decoded]
      properties:
        type:
          type: string
          description: Detected or provided XDR type
        decoded:
          type: object
          description: Human-readable JSON representation of the XDR
          additionalProperties: true

    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: Human-readable error message